+++ frontend/src/app/(app)/(root)/people/page.tsx
@@ -5,10 +5,12 @@ import Link from 'next/link';
 export default function People() {
   return (
     <div>
-      <h1 className="mb-10 text-lg font-semibold leading-6 text-navy-900">People</h1>
-      <Button>
-        <Link href={routes.people.addNew.personalDetails}>+ Employee</Link>
-      </Button>
+      <div className="mb-6 flex items-center justify-between">
+        <h1 className="text-lg font-semibold leading-6 text-navy-900">People</h1>
+        <Button>
+          <Link href={routes.people.addNew.personalDetails}>+ Employee</Link>
+        </Button>
+      </div>
     </div>
   );
 }
