+++ frontend/src/components/pages/PersonalDetails/index.ts
@@ -1,0 +1,1 @@
+export { PersonalDetails } from './PersonalDetails';
