+++ frontend/src/store/people/index.ts
@@ -1,0 +1,1 @@
+export { usePeopleStore } from './store';
